requirements.txt
pandas>=1.3.0
numpy>=1.21.0
matplotlib>=3.4.0
seaborn>=0.11.0
plotly>=5.0.0
wordcloud>=1.8.0
nltk>=3.6
scikit-learn>=1.0.0
xgboost>=1.5.0
tensorflow>=2.8.0
transformers>=4.15.0
Pillow>=8.3.0

